test: commit the probe capture-cost bench the figures came from
probe.md cited "median of seven warm builds" for figures produced by a throwaway harness that was deleted, so nothing in the repo could re-derive them. The harness is now bench/probe-capture-cost.test.ts beside the retrieval bench, sharing one extracted fixture so both cost docs quote the same pool. Re-measurement moved payload assembly (deep) from ~12/~25ms to ~8/~16ms; every other figure held. retrieval.md's table now says which of its rows the bench emits and which stay ad-hoc.
